+++ src/base/asset_loader_zip.c
@@ -31,6 +31,9 @@ typedef struct _asset_loader_zip_t {
   void* data;
   uint32_t size;
   mz_zip_archive archive;
+
+  data_reader_t* reader;
+  bool_t own_the_reader;
 } asset_loader_zip_t;
 
 #ifdef WITH_SDL
@@ -101,6 +104,15 @@ static asset_info_t* asset_loader_zip_load(asset_loader_t* loader, uint16_t type
 static ret_t asset_loader_zip_destroy(asset_loader_t* loader) {
   asset_loader_zip_t* zip = (asset_loader_zip_t*)loader;
   mz_zip_end(&(zip->archive));
+
+  if(zip->data != NULL) {
+    TKMEM_FREE(zip->data);
+  }
+
+  if(zip->own_the_reader && zip->reader != NULL) {
+    data_reader_destroy(zip->reader);
+  }
+
   TKMEM_FREE(loader);
 
   return RET_OK;
@@ -109,19 +121,26 @@ static ret_t asset_loader_zip_destroy(asset_loader_t* loader) {
 static const asset_loader_vtable_t s_asset_loader_zip_vtable = {
     .load = asset_loader_zip_load, .destroy = asset_loader_zip_destroy};
 
-asset_loader_t* asset_loader_zip_create(const char* zipfile) {
+static asset_loader_zip_t* asset_loader_zip_create_default(void) {
   asset_loader_zip_t* zip = TKMEM_ZALLOC(asset_loader_zip_t);
   return_value_if_fail(zip != NULL, NULL);
-
-  zip->data = load_asset(zipfile, &(zip->size));
-  zip->asset_loader.vt = &s_asset_loader_zip_vtable;
-  if (zip->data == NULL) {
-    TKMEM_FREE(zip);
-  }
 
   zip->archive.m_pFree = miniz_awtk_free_func;
   zip->archive.m_pAlloc = miniz_awtk_alloc_func;
   zip->archive.m_pRealloc = miniz_awtk_realloc_func;
+  zip->asset_loader.vt = &s_asset_loader_zip_vtable;
+
+  return zip;
+}
+
+asset_loader_t* asset_loader_zip_create(const char* zipfile) {
+  asset_loader_zip_t* zip = asset_loader_zip_create_default();
+  return_value_if_fail(zip != NULL, NULL);
+
+  zip->data = load_asset(zipfile, &(zip->size));
+  if (zip->data == NULL) {
+    TKMEM_FREE(zip);
+  }
 
   if (!mz_zip_reader_init_mem(&(zip->archive), zip->data, zip->size, 0)) {
     TKMEM_FREE(zip);
@@ -129,4 +148,31 @@ asset_loader_t* asset_loader_zip_create(const char* zipfile) {
 
   return (asset_loader_t*)zip;
 }
+
+static size_t asset_loader_zip_read(void *pOpaque, mz_uint64 offset, void *buff, size_t n) {
+  asset_loader_zip_t* zip = (asset_loader_zip_t*)pOpaque;
+
+  return data_reader_read(zip->reader, offset, buff, n);
+}
+
+asset_loader_t* asset_loader_zip_create_with_reader(data_reader_t* reader, bool_t own_the_reader) {
+  asset_loader_zip_t* zip = NULL;
+  uint32_t size = data_reader_get_size(reader);
+  return_value_if_fail(reader != NULL, NULL);
+
+  zip = asset_loader_zip_create_default();
+  return_value_if_fail(zip != NULL, NULL);
+
+  zip->reader = reader;
+  zip->own_the_reader = own_the_reader;
+  zip->archive.m_pIO_opaque = zip;
+  zip->archive.m_pRead= asset_loader_zip_read;
+
+  if (!mz_zip_reader_init(&(zip->archive), size, 0)) {
+    TKMEM_FREE(zip);
+  }
+  
+  return (asset_loader_t*)zip;
+}
+
 #endif /*WITH_ASSET_LOADER_ZIP*/