+++ src/base/asset_loader_zip.h
@@ -22,6 +22,7 @@
 #ifndef TK_ASSET_LOADER_ZIP_H
 #define TK_ASSET_LOADER_ZIP_H
 
+#include "tkc/data_reader.h"
 #include "base/asset_loader.h"
 
 BEGIN_C_DECLS
@@ -34,6 +35,16 @@ BEGIN_C_DECLS
  * @return {asset_loader_t*} 返回loader对象。
  */
 asset_loader_t* asset_loader_zip_create(const char* zipfile);
+
+/**
+ * @method asset_loader_zip_create_with_reader
+ * 创建zip资源加载器。
+ * @param {data_reader_t*} reader reader对象。
+ * @param {bool_t} own_the_reader 如果希望销毁loader时销毁reader，请设置为TRUE。
+ *
+ * @return {asset_loader_t*} 返回loader对象。
+ */
+asset_loader_t* asset_loader_zip_create_with_reader(data_reader_t* reader, bool_t own_the_reader);
 
 END_C_DECLS